git-svn-id: https://chibios.svn.sourceforge.net/svnroot/chibios/branches/stm8_cosmic_dev@1994 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/ports/RC/STM8/chtypes.h
@@ -18,10 +18,10 @@
 */
 
 /**
- * @file    STM8/chtypes.h
- * @brief   STM8 port system types.
+ * @file    RC/STM8/chtypes.h
+ * @brief   STM8 (Raisonance) port system types.
  *
- * @addtogroup STM8_CORE
+ * @addtogroup STM8_RAISONANCE_CORE
  * @{
  */
 
@@ -94,7 +94,7 @@ typedef uint16_t        systime_t;
 /**
  * @brief   Counter, recommended fastest signed.
  */
-typedef int16_t         cnt_t;
+typedef int8_t          cnt_t;
 
 /**
  * @brief   Inline function modifier.
@@ -103,7 +103,7 @@ typedef int16_t         cnt_t;
 
 /**
  * @brief   ROM constant modifier.
- * @note    Uses the custom "code" keyword in this port.
+ * @note    Uses the "const" keyword in this port.
  */
 #define ROMCONST code
